Add npm start command
Installing the project dependencies does not provide the
user with a global webpack command so the setup instructions
don't work if you do not already have webpack-cli installed
globally. This change adds an npm command that will run the
locally installed webpack so there are no additional setup
steps required.